Fix to evalTrappedPiecesSide

+++ engine/eval/evaluator.hpp
@@ -101,8 +101,8 @@ private:
     static void computeAttackData(AttackData data[2], const chess::Board& b, uint64_t occ) noexcept;
     static inline void ensureAttackData(AttackData data[2], const chess::Board& b, uint64_t occ) noexcept;
 
-    static inline uint64_t knightAttacksLookup(int sq, uint64_t) noexcept;
-    template<uint64_t (*AttackFn)(int, uint64_t), int64_t PinnedPenalty, int64_t LowMobPenalty>
+    static inline uint64_t knightAttacksLookup(uint8_t sq, uint64_t) noexcept;
+    template<uint64_t (*AttackFn)(uint8_t, uint64_t), int64_t PinnedPenalty, int64_t LowMobPenalty>
     static inline int64_t evalTrappedPiecesGeneric(uint64_t piecesBb, uint64_t occ, uint64_t mobilityMask, int sign) noexcept;
     static inline int64_t evalTrappedPiecesSide(const chess::Board& b, uint64_t occ, int side, int sign) noexcept;
